Merge: from maintenance-7.2 up to revision 3150 into trunk.

+++ build/modules/services-ejb3/src/main/java/ccc/persistence/FsCoreData.java
@@ -33,6 +33,7 @@ import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
+import java.io.OutputStream;
 
 import org.apache.log4j.Logger;
 
@@ -55,6 +56,7 @@ class FsCoreData
     private static final Logger LOG = Logger.getLogger(FsCoreData.class);
 
     private final File _root;
+
 
     /**
      * Constructor.
@@ -85,23 +87,29 @@ class FsCoreData
         LOG.debug("Created file store: "+_root.getAbsolutePath());
     }
 
+
     /** {@inheritDoc} */
     @Override
     public Data create(final InputStream dataStream, final long length) {
         final Data d = new Data();
         final File dir = mkdir(d);
         final File f = new File(dir, d.getId().toString());
+        if (f.exists()) {
+            throw new RuntimeException("File exists: "+f.getAbsolutePath());
+        }
 
         try {
-            final FileOutputStream fos = new FileOutputStream(f);
+            final OutputStream fos = createOutputStream(f);
 
             try {
                 IO.copy(dataStream, fos);
+                fos.flush();
+                fos.close();
+
             } catch (final Exception e) {
                 LOG.error("Error writing to file "+f.getAbsolutePath(), e);
                 throw e;
             } finally {
-                attemptClose(f, fos);
                 attemptClose(f, dataStream);
             }
             LOG.debug("Wrote data to file store: "+f.getAbsolutePath());
@@ -160,6 +168,7 @@ class FsCoreData
 
 
     private void attemptDelete(final File f) {
+        if (!f.exists()) { return; }
         final boolean deleted = f.delete();
         if (!deleted) {
             LOG.error("Failed to delete file "+f.getAbsolutePath());
@@ -185,5 +194,19 @@ class FsCoreData
         final String sep = File.separator;
         final File dir = new File(_root, c0+sep+c1+sep+c2);
         return dir;
+    }
+
+
+    /**
+     * Create an output stream on the specified file.
+     *
+     * @param f The file to output to.
+     *
+     * @return An output stream on the specified file.
+     *
+     * @throws FileNotFoundException If stream creation fails.
+     */
+    OutputStream createOutputStream(final File f) throws FileNotFoundException {
+        return new FileOutputStream(f);
     }
 }
